requirements.txt
torchdiffeq
geotorch
requests==2.32.3
numpy==1.26.3
Jinja2==3.1.3
timm==1.0.9
tqdm==4.66.5
pandas==2.2.3
git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e#egg=autoattack
gdown==5.1.0
pyyaml
matplotlib==3.9.2
wandb==0.18.1
scikit-learn==1.5.2
loguru==0.7.2
datasets==3.4.1
opencv-python==4.11.0.86
